% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytic_functions.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Compute log-likelihood for linear model}
\usage{
log_likelihood(design, beta, outcome, noise_var = 1)
}
\description{
Compute log-likelihood for linear model
}
