% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytic_functions.R
\name{numeric_grad}
\alias{numeric_grad}
\title{Compute numeric gradient of log-likelihood for linear model}
\usage{
numeric_grad(design, beta, outcome, noise_var = 1)
}
\description{
Compute numeric gradient of log-likelihood for linear model
}
